services/agent-workers/requirements.txt
temporalio>=1.5.0
openai>=1.12.0
python-dotenv>=1.0.1
pydantic>=2.6.1
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
respx>=0.21.1
psycopg2-binary>=2.9.9
pgvector>=0.2.4
